Create useApi mock
