+++ src/Hedwig/ClientApp/src/containers/Roster/Roster.test.tsx
@@ -45,81 +45,7 @@ const waitForUpdate = async (wrapper: any) => {
 	wrapper.update();
 };
 
-// https://github.com/facebook/jest/issues/5579
-jest.mock('../../hooks/useApi', () => {
-	const moment = require('moment');
-	return {
-		__esModule: true,
-		default: (callback: any, dependencies: any[]) => {
-			return callback(
-				{
-					apiOrganizationsOrgIdSitesIdGet: (params: any) => [
-						false,
-						null,
-						{
-							id: 1,
-							name: "Children's Adventure Center",
-							organizationId: 1,
-							enrollments: undefined,
-							organization: undefined,
-						},
-					],
-					apiOrganizationsOrgIdSitesSiteIdEnrollmentsGet: (params: any) => [
-						false,
-						null,
-						[
-							{
-								child: {
-									id: 1,
-									firstName: 'James',
-									middleName: 'Sirius',
-									lastName: 'Potter',
-									birthdate: '2015-04-28',
-									suffix: null,
-								},
-								entry: '2019-01-01',
-								exit: null,
-								fundings: [],
-								id: 1,
-								ageGroup: 'preschool',
-								siteId: 1,
-							},
-							{
-								child: {
-									id: 2,
-									firstName: 'Lily',
-									middleName: 'Luna',
-									lastName: 'Potter',
-									birthdate: '2016-12-12',
-									suffix: null,
-								},
-								entry: '2019-03-03',
-								exit: null,
-								fundings: [
-									{
-										entry: '2019-03-01',
-										exit: '2019-04-01',
-										source: 'CDC',
-										time: 'part',
-									},
-								],
-								id: 2,
-								ageGroup: 'preschool',
-							},
-						].filter(e => {
-							return (
-								(!e.entry ? true : moment(e.entry).isBefore(params.endDate)) &&
-								(!e.exit ? true : moment(e.exit).isAfter(moment(params.startDate)))
-							);
-						}),
-					],
-				},
-				[]
-			);
-		},
-	};
-});
-
+jest.mock('../../hooks/useApi');
 import useApi from '../../hooks/useApi';
 
 beforeAll(() => {
